feat: DSTEW-1621 split application service into smaller use case services
