--- data-access-service/src/main/java/uk/gov/justice/laa/dstew/access/service/ApplicationService.java
@@ -1,666 +1,0 @@
-package uk.gov.justice.laa.dstew.access.service;
-
-import jakarta.transaction.Transactional;
-import java.time.Instant;
-import java.util.ArrayList;
-import java.util.LinkedHashSet;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-import java.util.Set;
-import java.util.UUID;
-import java.util.stream.Collectors;
-import lombok.NonNull;
-import org.springframework.stereotype.Service;
-import tools.jackson.core.type.TypeReference;
-import tools.jackson.databind.ObjectMapper;
-import uk.gov.justice.laa.dstew.access.entity.ApplicationEntity;
-import uk.gov.justice.laa.dstew.access.entity.CaseworkerEntity;
-import uk.gov.justice.laa.dstew.access.entity.CertificateEntity;
-import uk.gov.justice.laa.dstew.access.entity.DecisionEntity;
-import uk.gov.justice.laa.dstew.access.entity.MeritsDecisionEntity;
-import uk.gov.justice.laa.dstew.access.entity.NoteEntity;
-import uk.gov.justice.laa.dstew.access.entity.ProceedingEntity;
-import uk.gov.justice.laa.dstew.access.exception.ResourceNotFoundException;
-import uk.gov.justice.laa.dstew.access.mapper.ApplicationMapper;
-import uk.gov.justice.laa.dstew.access.mapper.MapperUtil;
-import uk.gov.justice.laa.dstew.access.mapper.NoteMapper;
-import uk.gov.justice.laa.dstew.access.mapper.ProceedingMapper;
-import uk.gov.justice.laa.dstew.access.model.ApplicationContent;
-import uk.gov.justice.laa.dstew.access.model.ApplicationCreateRequest;
-import uk.gov.justice.laa.dstew.access.model.ApplicationMerits;
-import uk.gov.justice.laa.dstew.access.model.ApplicationNoteResponse;
-import uk.gov.justice.laa.dstew.access.model.ApplicationNotesResponse;
-import uk.gov.justice.laa.dstew.access.model.ApplicationProceedingResponse;
-import uk.gov.justice.laa.dstew.access.model.ApplicationResponse;
-import uk.gov.justice.laa.dstew.access.model.ApplicationUpdateRequest;
-import uk.gov.justice.laa.dstew.access.model.CreateNoteRequest;
-import uk.gov.justice.laa.dstew.access.model.DecisionStatus;
-import uk.gov.justice.laa.dstew.access.model.DomainEventType;
-import uk.gov.justice.laa.dstew.access.model.EventHistoryRequest;
-import uk.gov.justice.laa.dstew.access.model.LinkedApplication;
-import uk.gov.justice.laa.dstew.access.model.MakeDecisionProceedingRequest;
-import uk.gov.justice.laa.dstew.access.model.MakeDecisionRequest;
-import uk.gov.justice.laa.dstew.access.model.MeritsDecisionStatus;
-import uk.gov.justice.laa.dstew.access.repository.ApplicationRepository;
-import uk.gov.justice.laa.dstew.access.repository.CaseworkerRepository;
-import uk.gov.justice.laa.dstew.access.repository.CertificateRepository;
-import uk.gov.justice.laa.dstew.access.repository.DecisionRepository;
-import uk.gov.justice.laa.dstew.access.repository.MeritsDecisionRepository;
-import uk.gov.justice.laa.dstew.access.repository.NoteRepository;
-import uk.gov.justice.laa.dstew.access.repository.ProceedingRepository;
-import uk.gov.justice.laa.dstew.access.security.AllowApiCaseworker;
-import uk.gov.justice.laa.dstew.access.utils.VersionCheckHelper;
-import uk.gov.justice.laa.dstew.access.validation.ApplicationValidations;
-import uk.gov.justice.laa.dstew.access.validation.PayloadValidationService;
-import uk.gov.justice.laa.dstew.access.validation.ValidationException;
-
-/** Service class for managing Applications. */
-@Service
-public class ApplicationService {
-
-  private final int applicationVersion = 1;
-  private final ApplicationRepository applicationRepository;
-  private final ApplicationMapper applicationMapper;
-  private final ProceedingMapper proceedingMapper;
-  private final ApplicationValidations applicationValidations;
-  private final ObjectMapper objectMapper;
-  private final CaseworkerRepository caseworkerRepository;
-  private final DomainEventService domainEventService;
-  private final ApplicationContentParserService applicationContentParser;
-  private final DecisionRepository decisionRepository;
-  private final ProceedingRepository proceedingRepository;
-  private final MeritsDecisionRepository meritsDecisionRepository;
-  private final CertificateRepository certificateRepository;
-  private final ProceedingsService proceedingsService;
-  private final PayloadValidationService payloadValidationService;
-  private final NoteRepository noteRepository;
-  private final NoteMapper noteMapper;
-
-  /**
-   * Constructs an ApplicationService with required dependencies.
-   *
-   * @param applicationRepository the repository
-   * @param applicationMapper the mapper between entity and DTO
-   * @param applicationValidations validations for requests
-   * @param objectMapper Jackson ObjectMapper for JSONB
-   */
-  public ApplicationService(
-      final ApplicationRepository applicationRepository,
-      final ApplicationMapper applicationMapper,
-      final ProceedingMapper proceedingMapper,
-      final ApplicationValidations applicationValidations,
-      final ObjectMapper objectMapper,
-      final CaseworkerRepository caseworkerRepository,
-      final DecisionRepository decisionRepository,
-      final DomainEventService domainEventService,
-      final ApplicationContentParserService applicationContentParserService,
-      final ProceedingRepository proceedingRepository,
-      final MeritsDecisionRepository meritsDecisionRepository,
-      final CertificateRepository certificateRepository,
-      final ProceedingsService proceedingsService,
-      final PayloadValidationService payloadValidationService,
-      final NoteRepository noteRepository,
-      final NoteMapper noteMapper) {
-    this.applicationRepository = applicationRepository;
-    this.applicationMapper = applicationMapper;
-    this.proceedingMapper = proceedingMapper;
-    this.applicationValidations = applicationValidations;
-    this.applicationContentParser = applicationContentParserService;
-    this.proceedingRepository = proceedingRepository;
-    this.proceedingsService = proceedingsService;
-    this.payloadValidationService = payloadValidationService;
-    this.objectMapper = objectMapper;
-    this.caseworkerRepository = caseworkerRepository;
-    this.domainEventService = domainEventService;
-    this.decisionRepository = decisionRepository;
-    this.meritsDecisionRepository = meritsDecisionRepository;
-    this.certificateRepository = certificateRepository;
-    this.noteRepository = noteRepository;
-    this.noteMapper = noteMapper;
-  }
-
-  /**
-   * Retrieve a single application by ID.
-   *
-   * @param id application UUID
-   * @return application DTO
-   */
-  @AllowApiCaseworker
-  public ApplicationResponse getApplication(final UUID id) {
-    final ApplicationEntity entity = checkIfApplicationExists(id);
-    ApplicationResponse application = applicationMapper.toApplication(entity);
-
-    Set<ProceedingEntity> proceedings = proceedingRepository.findAllByApplicationId(id);
-
-    if (proceedings != null) {
-
-      proceedings.forEach(
-          proceeding -> {
-            ApplicationProceedingResponse applicationProceedingResponse =
-                proceedingMapper.toApplicationProceeding(proceeding);
-
-            // List<Map<String, Object>> involvedChildren = getInvolvedChildren(entity);
-            // if (involvedChildren != null) {
-            //   List<Object> children = new ArrayList<>();
-            //   involvedChildren.forEach(children::add);
-            //   applicationProceeding.setInvolvedChildren(children);
-            // }
-            // else {
-            //   applicationProceeding.setInvolvedChildren(null);
-            // }
-
-            if (entity.getDecision() != null) {
-              Optional<MeritsDecisionEntity> meritsDecision =
-                  entity.getDecision().getMeritsDecisions().stream()
-                      .filter(m -> m.getProceeding().getId() == proceeding.getId())
-                      .findFirst();
-
-              meritsDecision.ifPresent(
-                  meritsDecisionEntity ->
-                      applicationProceedingResponse.setMeritsDecision(
-                          meritsDecisionEntity.getDecision()));
-            }
-            application.getProceedings().add(applicationProceedingResponse);
-          });
-    }
-
-    return application;
-  }
-
-  private List<Map<String, Object>> getInvolvedChildren(ApplicationEntity entity) {
-
-    ApplicationContent applicationContent =
-        MapperUtil.getObjectMapper()
-            .convertValue(entity.getApplicationContent(), ApplicationContent.class);
-    ApplicationMerits meritsObj = applicationContent.getApplicationMerits();
-
-    if (meritsObj == null) {
-      return null;
-    }
-
-    return meritsObj.getInvolvedChildren();
-  }
-
-  /**
-   * Create a new application.
-   *
-   * @param req DTO containing creation fields
-   * @return UUID of the created application
-   */
-  @AllowApiCaseworker
-  @Transactional
-  public UUID createApplication(final ApplicationCreateRequest req) {
-    ApplicationEntity entity = applicationMapper.toApplicationEntity(req);
-    ApplicationContent applicationContent =
-        payloadValidationService.convertAndValidate(
-            req.getApplicationContent(), ApplicationContent.class);
-    setValuesFromApplicationContent(entity, applicationContent);
-    checkForDuplicateApplication(entity.getApplyApplicationId());
-    entity.setSchemaVersion(applicationVersion);
-
-    final ApplicationEntity saved = applicationRepository.save(entity);
-
-    linkToLeadApplicationIfApplicable(applicationContent, saved);
-    proceedingsService.saveProceedings(applicationContent, saved.getId());
-    domainEventService.saveCreateApplicationDomainEvent(saved, req, null);
-    createAndSendHistoricRecord(saved, null);
-
-    return saved.getId();
-  }
-
-  /**
-   * Sets key fields in the application entity based on parsed application content.
-   *
-   * @param entity application entity to update
-   * @param requestAppContent application content from the request
-   */
-  private void setValuesFromApplicationContent(
-      ApplicationEntity entity, ApplicationContent requestAppContent) {
-
-    var parsedContentDetails =
-        applicationContentParser.normaliseApplicationContentDetails(requestAppContent);
-    entity.setApplyApplicationId(parsedContentDetails.applyApplicationId());
-    entity.setUsedDelegatedFunctions(parsedContentDetails.usedDelegatedFunctions());
-    entity.setCategoryOfLaw(parsedContentDetails.categoryOfLaw());
-    entity.setMatterType(parsedContentDetails.matterType());
-    entity.setSubmittedAt(parsedContentDetails.submittedAt());
-    entity.setOfficeCode(parsedContentDetails.officeCode());
-  }
-
-  /**
-   * Validates that the application ID is not a duplicate.
-   *
-   * @param applyApplicationId the UUID of the application to check
-   * @throws ValidationException if a duplicate application exists
-   */
-  private void checkForDuplicateApplication(final UUID applyApplicationId) {
-    if (applicationRepository.existsByApplyApplicationId(applyApplicationId)) {
-      throw new ValidationException(
-          List.of("Application already exists for Apply Application Id: " + applyApplicationId));
-    }
-  }
-
-  /**
-   * Update an existing application.
-   *
-   * @param id application UUID
-   * @param req DTO with update fields
-   */
-  @AllowApiCaseworker
-  public void updateApplication(final UUID id, final ApplicationUpdateRequest req) {
-    final ApplicationEntity entity = checkIfApplicationExists(id);
-    applicationValidations.checkApplicationUpdateRequest(req);
-    applicationMapper.updateApplicationEntity(entity, req);
-    entity.setModifiedAt(Instant.now());
-    applicationRepository.save(entity);
-
-    domainEventService.saveUpdateApplicationDomainEvent(entity, null);
-
-    // Optional: create snapshot for audit/history
-    objectMapper.convertValue(
-        applicationMapper.toApplication(entity), new TypeReference<Map<String, Object>>() {});
-  }
-
-  /**
-   * Create a note for an application.
-   *
-   * @param id UUID of the application
-   * @param request note to be created
-   */
-  @AllowApiCaseworker
-  @Transactional
-  public void createApplicationNote(final UUID id, final CreateNoteRequest request) {
-    ApplicationEntity application = checkIfApplicationExists(id);
-    noteRepository.save(NoteEntity.builder().applicationId(id).notes(request.getNotes()).build());
-    domainEventService.saveCreateApplicationNoteDomainEvent(application, request);
-  }
-
-  /**
-   * Retrieve all notes for an application in ascending date order.
-   *
-   * @param id UUID of the application
-   * @return response containing list of notes in ascending createdAt order
-   */
-  @AllowApiCaseworker
-  public ApplicationNotesResponse getApplicationNotes(final UUID id) {
-    checkIfApplicationExists(id);
-    List<ApplicationNoteResponse> notes =
-        noteRepository.findByApplicationIdOrderByCreatedAtAsc(id).stream()
-            .map(noteMapper::toApplicationNoteResponse)
-            .toList();
-    return new ApplicationNotesResponse().notes(notes);
-  }
-
-  /**
-   * Placeholder for historic/audit record creation.
-   *
-   * @param entity application entity
-   * @param actionType optional action type
-   */
-  protected void createAndSendHistoricRecord(
-      final ApplicationEntity entity, final Object actionType) {
-    // Implement audit/history publishing if required
-  }
-
-  /**
-   * Check existence of an application by ID.
-   *
-   * @param id UUID of application
-   * @return found entity
-   */
-  private ApplicationEntity checkIfApplicationExists(final UUID id) {
-    return applicationRepository
-        .findById(id)
-        .orElseThrow(
-            () ->
-                new ResourceNotFoundException(
-                    String.format("No application found with id: %s", id)));
-  }
-
-  /**
-   * Check existence of a caseworker by ID.
-   *
-   * @param caseworkerId userid of caseworker
-   * @return found entity
-   */
-  private CaseworkerEntity checkIfCaseworkerExists(final UUID caseworkerId) {
-    return caseworkerRepository
-        .findById(caseworkerId)
-        .orElseThrow(
-            () ->
-                new ResourceNotFoundException(
-                    String.format("No caseworker found with id: %s", caseworkerId)));
-  }
-
-  /**
-   * Checks that applications exist for all the IDs provided.
-   *
-   * @param ids Collection of UUIDs of applications
-   * @return found entity
-   */
-  private List<ApplicationEntity> checkIfAllApplicationsExist(final List<UUID> ids) {
-    applicationValidations.checkApplicationIdList(ids);
-    var idsToFetch = ids.stream().distinct().toList();
-    var applications = applicationRepository.findAllById(idsToFetch);
-    List<UUID> fetchedApplicationsIds =
-        applications.stream().map(ApplicationEntity::getId).toList();
-    String missingIds =
-        idsToFetch.stream()
-            .filter(appId -> !fetchedApplicationsIds.contains(appId))
-            .map(UUID::toString)
-            .collect(Collectors.joining(","));
-    if (!missingIds.isEmpty()) {
-      String exceptionMsg = "No application found with ids: " + missingIds;
-      throw new ResourceNotFoundException(exceptionMsg);
-    }
-    return applications;
-  }
-
-  /**
-   * Checks that applications exist for all the IDs provided.
-   *
-   * @param associatedApplyIds Collection of apply applications ids
-   */
-  private void checkIfAllAssociatedApplicationsExist(final List<UUID> associatedApplyIds) {
-    applicationValidations.checkApplicationIdList(associatedApplyIds);
-    List<UUID> foundApplyAppIds =
-        applicationRepository.findAllByApplyApplicationIdIn(associatedApplyIds).stream()
-            .map(ApplicationEntity::getApplyApplicationId)
-            .toList();
-    if (foundApplyAppIds.size() != associatedApplyIds.size()) {
-      List<UUID> remainingIds =
-          associatedApplyIds.stream().filter(id -> !foundApplyAppIds.contains(id)).toList();
-      String exceptionMsg =
-          "No linked application found with associated apply ids: " + remainingIds;
-      throw new ResourceNotFoundException(exceptionMsg);
-    }
-  }
-
-  /**
-   * Assigns a caseworker to an application.
-   *
-   * @param caseworkerId the UUID of the caseworker to assign
-   * @param applicationIds the UUIDs of the applications to assign the caseworker to
-   * @throws ResourceNotFoundException if the application or caseworker does not exist
-   */
-  @Transactional
-  @AllowApiCaseworker
-  public void assignCaseworker(
-      @NonNull final UUID caseworkerId,
-      final List<UUID> applicationIds,
-      final EventHistoryRequest eventHistoryRequest) {
-    final CaseworkerEntity caseworker = checkIfCaseworkerExists(caseworkerId);
-
-    final List<ApplicationEntity> applications = checkIfAllApplicationsExist(applicationIds);
-
-    applications.forEach(
-        app -> {
-          if (!applicationCurrentCaseworkerIsCaseworker(app, caseworker)) {
-            app.setCaseworker(caseworker);
-            app.setModifiedAt(Instant.now());
-            applicationRepository.save(app);
-          }
-
-          domainEventService.saveAssignApplicationDomainEvent(
-              app.getId(), caseworker.getId(), eventHistoryRequest.getEventDescription());
-        });
-  }
-
-  /**
-   * Unassigns a caseworker from an application.
-   *
-   * @param applicationId the UUID of the application to update
-   * @throws ResourceNotFoundException if the application does not exist
-   */
-  @AllowApiCaseworker
-  public void unassignCaseworker(final UUID applicationId, EventHistoryRequest history) {
-    final ApplicationEntity entity = checkIfApplicationExists(applicationId);
-
-    if (entity.getCaseworker() == null) {
-      return;
-    }
-
-    entity.setCaseworker(null);
-    entity.setModifiedAt(Instant.now());
-
-    applicationRepository.save(entity);
-
-    domainEventService.saveUnassignApplicationDomainEvent(
-        entity.getId(), null, history.getEventDescription());
-  }
-
-  /**
-   * Check if an application has a caseworker assigned already and checks if the assigned caseworker
-   * matches the given caseworker.
-   */
-  private static boolean applicationCurrentCaseworkerIsCaseworker(
-      ApplicationEntity application, CaseworkerEntity caseworker) {
-    return application.getCaseworker() != null && application.getCaseworker().equals(caseworker);
-  }
-
-  /**
-   * Update an existing application to add the decision details.
-   *
-   * @param applicationId application UUID
-   * @param request DTO with update fields
-   */
-  @Transactional
-  @AllowApiCaseworker
-  public void makeDecision(final UUID applicationId, final MakeDecisionRequest request) {
-    final ApplicationEntity application = checkIfApplicationExists(applicationId);
-    VersionCheckHelper.checkEntityVersionLocking(
-        applicationId, application.getVersion(), request.getApplicationVersion());
-    final UUID caseworkerId = getCaseworkerId(applicationId, application);
-    applicationValidations.checkApplicationMakeDecisionRequest(request);
-
-    application.setModifiedAt(Instant.now());
-    application.setIsAutoGranted(request.getAutoGranted());
-    applicationRepository.save(application);
-
-    DecisionEntity decision =
-        application.getDecision() != null
-            ? application.getDecision()
-            : DecisionEntity.builder().meritsDecisions(Set.of()).build();
-
-    Set<MeritsDecisionEntity> merits = new LinkedHashSet<>(decision.getMeritsDecisions());
-
-    Map<UUID, ProceedingEntity> proceedingEntityMap =
-        getUuidProceedingEntityMap(applicationId, request);
-
-    request
-        .getProceedings()
-        .forEach(
-            proceeding -> {
-              ProceedingEntity proceedingEntity =
-                  proceedingEntityMap.get(proceeding.getProceedingId());
-              saveMeritsDecisionEntity(proceeding, decision, proceedingEntity, merits);
-            });
-
-    decision.setMeritsDecisions(merits);
-    decision.setOverallDecision(DecisionStatus.valueOf(request.getOverallDecision().getValue()));
-    decision.setModifiedAt(Instant.now());
-    decisionRepository.save(decision);
-
-    // Persist certificate if overallDecision is GRANTED
-    if (decision.getOverallDecision() == DecisionStatus.GRANTED
-        && request.getCertificate() != null) {
-      CertificateEntity certificate =
-          certificateRepository
-              .findByApplicationId(applicationId)
-              .map(
-                  existing -> {
-                    existing.setCertificateContent(request.getCertificate());
-                    return existing;
-                  })
-              .orElseGet(
-                  () ->
-                      CertificateEntity.builder()
-                          .applicationId(applicationId)
-                          .certificateContent(request.getCertificate())
-                          .build());
-
-      certificateRepository.save(certificate);
-    }
-
-    if (decision.getOverallDecision() == DecisionStatus.REFUSED) {
-      if (certificateRepository.existsByApplicationId(applicationId)) {
-        certificateRepository.deleteByApplicationId(applicationId);
-      }
-    }
-
-    switch (decision.getOverallDecision()) {
-      case GRANTED ->
-          domainEventService.saveMakeDecisionDomainEvent(
-              applicationId,
-              request,
-              caseworkerId,
-              DomainEventType.APPLICATION_MAKE_DECISION_GRANTED);
-      case REFUSED ->
-          domainEventService.saveMakeDecisionDomainEvent(
-              applicationId,
-              request,
-              caseworkerId,
-              DomainEventType.APPLICATION_MAKE_DECISION_REFUSED);
-      default ->
-          throw new IllegalStateException("Unexpected value: " + decision.getOverallDecision());
-    }
-
-    if (application.getDecision() == null) {
-      application.setDecision(decision);
-      applicationRepository.save(application);
-    }
-  }
-
-  private void saveMeritsDecisionEntity(
-      MakeDecisionProceedingRequest proceeding,
-      DecisionEntity decision,
-      ProceedingEntity proceedingEntity,
-      Set<MeritsDecisionEntity> merits) {
-    MeritsDecisionEntity meritDecisionEntity =
-        decision.getMeritsDecisions().stream()
-            .filter(m -> m.getProceeding().getId().equals(proceeding.getProceedingId()))
-            .findFirst()
-            .orElseGet(
-                () -> {
-                  MeritsDecisionEntity newEntity = new MeritsDecisionEntity();
-                  newEntity.setProceeding(proceedingEntity);
-                  return newEntity;
-                });
-
-    meritDecisionEntity.setModifiedAt(Instant.now());
-    meritDecisionEntity.setDecision(
-        MeritsDecisionStatus.valueOf(proceeding.getMeritsDecision().getDecision().toString()));
-    meritDecisionEntity.setReason(proceeding.getMeritsDecision().getReason());
-    meritDecisionEntity.setJustification(proceeding.getMeritsDecision().getJustification());
-    meritsDecisionRepository.save(meritDecisionEntity);
-    merits.add(meritDecisionEntity);
-  }
-
-  private @NonNull Map<UUID, ProceedingEntity> getUuidProceedingEntityMap(
-      UUID applicationId, MakeDecisionRequest request) {
-    List<UUID> proceedingIds =
-        request.getProceedings().stream()
-            .map(MakeDecisionProceedingRequest::getProceedingId)
-            .toList();
-    List<ProceedingEntity> proceedingEntities =
-        checkIfAllProceedingsExistForApplication(applicationId, proceedingIds);
-    return proceedingEntities.stream()
-        .collect(Collectors.toMap(ProceedingEntity::getId, proceeding -> proceeding));
-  }
-
-  private static UUID getCaseworkerId(UUID applicationId, ApplicationEntity application) {
-    final CaseworkerEntity caseworker = application.getCaseworker();
-    // This logic will be implemented in the next iteration when security is implemented in the
-    // service
-    //    if (caseworker == null) {
-    //      throw new ResourceNotFoundException(
-    //          String.format("Caseworker not found for application id: %s", applicationId)
-    //      );
-    //    }
-    return caseworker != null ? caseworker.getId() : null;
-  }
-
-  /**
-   * Checks that all provided proceeding IDs exist and are linked to the specified application.
-   * Throws a {@link ResourceNotFoundException} if any proceeding does not exist or is not linked to
-   * the given application.
-   *
-   * @param applicationId the UUID of the application to check proceedings against
-   * @param proceedingIds the list of proceeding UUIDs to validate
-   * @return a list of {@link ProceedingEntity} objects corresponding to the provided IDs
-   * @throws ResourceNotFoundException if any proceeding is missing or not linked to the application
-   */
-  private List<ProceedingEntity> checkIfAllProceedingsExistForApplication(
-      final UUID applicationId, final List<UUID> proceedingIds) {
-    List<UUID> idsToFetch = proceedingIds.stream().distinct().toList();
-    List<ProceedingEntity> proceedings = proceedingRepository.findAllById(idsToFetch);
-
-    List<UUID> foundProceedingIds = proceedings.stream().map(ProceedingEntity::getId).toList();
-
-    String proceedingIdsNotFound =
-        idsToFetch.stream()
-            .filter(id -> !foundProceedingIds.contains(id))
-            .map(UUID::toString)
-            .collect(Collectors.joining(","));
-
-    String proceedingIdsNotLinkedToApplication =
-        proceedings.stream()
-            .filter(p -> !p.getApplicationId().equals(applicationId))
-            .map(p -> p.getId().toString())
-            .collect(Collectors.joining(","));
-
-    if (!proceedingIdsNotFound.isEmpty() || !proceedingIdsNotLinkedToApplication.isEmpty()) {
-      List<String> errors = new ArrayList<>();
-      if (!proceedingIdsNotFound.isEmpty()) {
-        errors.add("No proceeding found with id: " + proceedingIdsNotFound);
-      }
-      if (!proceedingIdsNotLinkedToApplication.isEmpty()) {
-        errors.add("Not linked to application: " + proceedingIdsNotLinkedToApplication);
-      }
-      throw new ResourceNotFoundException(String.join("; ", errors));
-    }
-    return proceedings;
-  }
-
-  private static UUID getLeadApplicationId(List<LinkedApplication> linkedApplications) {
-    return (linkedApplications != null && !linkedApplications.isEmpty())
-        ? linkedApplications.getFirst().getLeadApplicationId()
-        : null;
-  }
-
-  private Optional<ApplicationEntity> getLeadApplication(ApplicationContent requestContent) {
-    final UUID leadApplicationId = getLeadApplicationId(requestContent.getAllLinkedApplications());
-    if (requestContent.getAllLinkedApplications() != null) {
-      List<UUID> list =
-          requestContent.getAllLinkedApplications().stream()
-              .map(LinkedApplication::getAssociatedApplicationId)
-              .filter(uuid -> !uuid.equals(requestContent.getId()))
-              .toList();
-
-      checkIfAllAssociatedApplicationsExist(list);
-    }
-
-    if (leadApplicationId == null) {
-      return Optional.empty();
-    }
-
-    var leadApplication = applicationRepository.findByApplyApplicationId(leadApplicationId);
-    if (leadApplication == null) {
-      throw new ResourceNotFoundException(
-          "Linking failed > Lead application not found, ID: " + leadApplicationId);
-    }
-
-    return Optional.of(leadApplication);
-  }
-
-  private void linkToLeadApplicationIfApplicable(
-      ApplicationContent appContent, ApplicationEntity entityToAdd) {
-    final Optional<ApplicationEntity> leadApplication = getLeadApplication(appContent);
-    leadApplication.ifPresent(
-        leadApp -> {
-          leadApp.addLinkedApplication(entityToAdd);
-          applicationRepository.save(leadApp);
-        });
-  }
-}
