+++ data-access-service/src/main/java/uk/gov/justice/laa/dstew/access/service/caseworker/AssignCaseworkerService.java
@@ -1,0 +1,107 @@
+package uk.gov.justice.laa.dstew.access.service.caseworker;
+
+import jakarta.transaction.Transactional;
+import java.time.Instant;
+import java.util.List;
+import java.util.UUID;
+import java.util.stream.Collectors;
+import lombok.NonNull;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import uk.gov.justice.laa.dstew.access.entity.ApplicationEntity;
+import uk.gov.justice.laa.dstew.access.entity.CaseworkerEntity;
+import uk.gov.justice.laa.dstew.access.exception.ResourceNotFoundException;
+import uk.gov.justice.laa.dstew.access.model.EventHistoryRequest;
+import uk.gov.justice.laa.dstew.access.repository.ApplicationRepository;
+import uk.gov.justice.laa.dstew.access.security.AllowApiCaseworker;
+import uk.gov.justice.laa.dstew.access.service.DomainEventService;
+import uk.gov.justice.laa.dstew.access.service.common.ServiceUtilities;
+import uk.gov.justice.laa.dstew.access.validation.ApplicationValidations;
+
+/** Service for working with caseworkers. */
+@RequiredArgsConstructor
+@Service
+public class AssignCaseworkerService {
+
+  private final ServiceUtilities serviceUtilities;
+  private final ApplicationRepository applicationRepository;
+  private final DomainEventService domainEventService;
+  final ApplicationValidations applicationValidations;
+
+  /**
+   * Assigns a caseworker to an application.
+   *
+   * @param caseworkerId the UUID of the caseworker to assign
+   * @param applicationIds the UUIDs of the applications to assign the caseworker to
+   * @throws ResourceNotFoundException if the application or caseworker does not exist
+   */
+  @Transactional
+  @AllowApiCaseworker
+  public void assignCaseworker(
+      @NonNull final UUID caseworkerId,
+      final List<UUID> applicationIds,
+      final EventHistoryRequest eventHistoryRequest) {
+    final CaseworkerEntity caseworker = serviceUtilities.checkIfCaseworkerExists(caseworkerId);
+
+    final List<ApplicationEntity> applications = checkIfAllApplicationsExist(applicationIds);
+
+    applications.forEach(
+        app -> {
+          if (!serviceUtilities.applicationCurrentCaseworkerIsCaseworker(app, caseworker)) {
+            app.setCaseworker(caseworker);
+            app.setModifiedAt(Instant.now());
+            applicationRepository.save(app);
+          }
+
+          domainEventService.saveAssignApplicationDomainEvent(
+              app.getId(), caseworker.getId(), eventHistoryRequest.getEventDescription());
+        });
+  }
+
+  /**
+   * Unassigns a caseworker from an application.
+   *
+   * @param applicationId the UUID of the application to update
+   * @throws ResourceNotFoundException if the application does not exist
+   */
+  @AllowApiCaseworker
+  public void unassignCaseworker(final UUID applicationId, EventHistoryRequest history) {
+    final ApplicationEntity entity = serviceUtilities.checkIfApplicationExists(applicationId);
+
+    if (entity.getCaseworker() == null) {
+      return;
+    }
+
+    entity.setCaseworker(null);
+    entity.setModifiedAt(Instant.now());
+
+    applicationRepository.save(entity);
+
+    domainEventService.saveUnassignApplicationDomainEvent(
+        entity.getId(), null, history.getEventDescription());
+  }
+
+  /**
+   * Checks that applications exist for all the IDs provided.
+   *
+   * @param ids Collection of UUIDs of applications
+   * @return found entity
+   */
+  private List<ApplicationEntity> checkIfAllApplicationsExist(final List<UUID> ids) {
+    applicationValidations.checkApplicationIdList(ids);
+    var idsToFetch = ids.stream().distinct().toList();
+    var applications = applicationRepository.findAllById(idsToFetch);
+    List<UUID> fetchedApplicationsIds =
+        applications.stream().map(ApplicationEntity::getId).toList();
+    String missingIds =
+        idsToFetch.stream()
+            .filter(appId -> !fetchedApplicationsIds.contains(appId))
+            .map(UUID::toString)
+            .collect(Collectors.joining(","));
+    if (!missingIds.isEmpty()) {
+      String exceptionMsg = "No application found with ids: " + missingIds;
+      throw new ResourceNotFoundException(exceptionMsg);
+    }
+    return applications;
+  }
+}
