+++ data-access-service/src/main/java/uk/gov/justice/laa/dstew/access/service/notes/GetNotesService.java
@@ -1,0 +1,38 @@
+package uk.gov.justice.laa.dstew.access.service.notes;
+
+import java.util.List;
+import java.util.UUID;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import uk.gov.justice.laa.dstew.access.mapper.NoteMapper;
+import uk.gov.justice.laa.dstew.access.model.ApplicationNoteResponse;
+import uk.gov.justice.laa.dstew.access.model.ApplicationNotesResponse;
+import uk.gov.justice.laa.dstew.access.repository.NoteRepository;
+import uk.gov.justice.laa.dstew.access.security.AllowApiCaseworker;
+import uk.gov.justice.laa.dstew.access.service.common.ServiceUtilities;
+
+/** Get application notes service. */
+@RequiredArgsConstructor
+@Service
+public class GetNotesService {
+
+  private final ServiceUtilities serviceUtilities;
+  private final NoteRepository noteRepository;
+  private final NoteMapper noteMapper;
+
+  /**
+   * Retrieve all notes for an application in ascending date order.
+   *
+   * @param id UUID of the application
+   * @return response containing list of notes in ascending createdAt order
+   */
+  @AllowApiCaseworker
+  public ApplicationNotesResponse getApplicationNotes(final UUID id) {
+    serviceUtilities.checkIfApplicationExists(id);
+    List<ApplicationNoteResponse> notes =
+        noteRepository.findByApplicationIdOrderByCreatedAtAsc(id).stream()
+            .map(noteMapper::toApplicationNoteResponse)
+            .toList();
+    return new ApplicationNotesResponse().notes(notes);
+  }
+}
