+++ data-access-service/src/test/java/uk/gov/justice/laa/dstew/access/service/application/CreateNoteTest.java
@@ -26,7 +26,7 @@ import uk.gov.justice.laa.dstew.access.exception.ResourceNotFoundException;
 import uk.gov.justice.laa.dstew.access.model.CreateNoteRequest;
 import uk.gov.justice.laa.dstew.access.model.DomainEventType;
 import uk.gov.justice.laa.dstew.access.model.ServiceName;
-import uk.gov.justice.laa.dstew.access.service.ApplicationService;
+import uk.gov.justice.laa.dstew.access.service.notes.CreateNoteService;
 import uk.gov.justice.laa.dstew.access.utils.BaseServiceTest;
 import uk.gov.justice.laa.dstew.access.utils.TestConstants;
 import uk.gov.justice.laa.dstew.access.utils.generator.DataGenerator;
@@ -36,7 +36,7 @@ import uk.gov.justice.laa.dstew.access.utils.generator.notes.CreateNoteRequestGe
 
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 public class CreateNoteTest extends BaseServiceTest {
-  @Autowired private ApplicationService serviceUnderTest;
+  @Autowired private CreateNoteService serviceUnderTest;
 
   @Autowired private ObjectMapper objectMapper;
